refactor: use u32 instead of u8 and usize
In frame buffer I used u8 and usize a lot and it didn't make any sense.
So, I used u32 for all instead.

+++ src/emulation/frame_buffer.cpp
@@ -2,7 +2,7 @@
 
 namespace shvav8 {
 
-bool FrameBuffer::draw(const u8 x, const u8 y) {
+bool FrameBuffer::draw(const u32 x, const u32 y) {
     const usize i = index(x % FrameBuffer::COLUMNS, y % FrameBuffer::ROWS);
     const bool collision = m_frame_buffer[i];
     m_frame_buffer[i] ^= 1;
@@ -13,8 +13,8 @@ void FrameBuffer::clear() { m_frame_buffer.fill(0); }
 
 std::vector<u32> FrameBuffer::get_drawn_pixels() const {
     std::vector<u32> squares;
-    for (u8 row = 0; row < FrameBuffer::ROWS; ++row) {
-        for (u8 column = 0; column < FrameBuffer::COLUMNS; ++column) {
+    for (u32 row = 0; row < FrameBuffer::ROWS; ++row) {
+        for (u32 column = 0; column < FrameBuffer::COLUMNS; ++column) {
             if (is_drawn(column, row)) {
                 squares.push_back(index(column, row));
             }
@@ -23,13 +23,13 @@ std::vector<u32> FrameBuffer::get_drawn_pixels() const {
     return squares;
 }
 
-bool FrameBuffer::is_drawn(const u8 x, const u8 y) const { return m_frame_buffer[index(x, y)]; }
+bool FrameBuffer::is_drawn(const u32 x, const u32 y) const { return m_frame_buffer[index(x, y)]; }
 
-usize FrameBuffer::index(const u8 x, const u8 y) const { return y * FrameBuffer::COLUMNS + x; }
+u32 FrameBuffer::index(const u32 x, const u32 y) const { return y * FrameBuffer::COLUMNS + x; }
 
 std::ostream& operator<<(std::ostream& ostream, const FrameBuffer& display) {
-    for (u8 y = 0; y < FrameBuffer::ROWS; ++y) {
-        for (u8 x = 0; x < FrameBuffer::COLUMNS; ++x) {
+    for (u32 y = 0; y < FrameBuffer::ROWS; ++y) {
+        for (u32 x = 0; x < FrameBuffer::COLUMNS; ++x) {
             if (display.is_drawn(x, y)) {
                 ostream << "* ";
             } else {
